All in one file
I have put all the fcc code in `lc8_reader.py` for simplicity.
There is now a class to hopefully do all the dirty work, but I
think I need to make sure that the input data are sane, the
definition of `rho_burn` is also consistent, and check what is
going on. Currently, I am ignoring Band 1 (which I think it's the
deep blue band), but the results are very strange indeed...