Cleaning up of LC8 reader
The LC8 reader has been cleaned up. This file only defines the
main class for the reader, as well as an exception when the
passed file(s) aren't LC8 format. I have also changed the data
class to sture the surface reflectance VRT file directly, as well
as the mask.