Further SIAC refinements
Codebase now works well with SIAC datasets. In this commit, the
software should be able to
(i) Recognise SIAC or L2A (e.g. Sen2Cor) files
(ii) Properly apply the SIAC cloud mask (I think). This requires
a threshold, which I have set to 10, but a different value might
be better (5 drops a lot of "good pixels", 20 adds a lot of cloud
or shadows).

Now testing on data...