Allow user to fixe a0 and a1
This change allows the user to specify
values for a0 and a1 for processing.
This means that we only solve for fcc
under the assumption of those parameters
being fixed. Appears to work, and some
modulation in fcc bleeds through as
possibly topographic effects not
partially correted by fitting a0/a1 directly.